refactor: Update MoviesPage to handle search functionality
This commit updates the MoviesPage component to handle search functionality. It introduces a new state variable, isSearching, to track whether the user is currently searching for movies. When isSearching is true, the component displays the search results using the MovingGrid component. Otherwise, it displays the trending movies and other sections using the HorizontalScroll component. This enhancement improves the user experience by allowing users to easily search for movies and view the results.